chore: refactor most theme related classes into a /theme folder
stream_chat_theme.dart remains in /src because I feel it will cause too much of a mess if moved.
