+++ packages/stream_chat_flutter/lib/src/theme/channel_preview_theme.dart
@@ -1,0 +1,64 @@
+import 'package:flutter/material.dart';
+import 'package:stream_chat_flutter/src/theme/avatar_theme.dart';
+
+/// Theme for channel preview
+class ChannelPreviewTheme {
+  /// Constructor for creating [ChannelPreviewTheme]
+  const ChannelPreviewTheme({
+    this.title,
+    this.subtitle,
+    this.lastMessageAt,
+    this.avatarTheme,
+    this.unreadCounterColor,
+    this.indicatorIconSize,
+  });
+
+  /// Theme for title
+  final TextStyle? title;
+
+  /// Theme for subtitle
+  final TextStyle? subtitle;
+
+  /// Theme of last message at
+  final TextStyle? lastMessageAt;
+
+  /// Avatar theme
+  final AvatarTheme? avatarTheme;
+
+  /// Unread counter color
+  final Color? unreadCounterColor;
+
+  /// Indicator icon size
+  final double? indicatorIconSize;
+
+  /// Copy with theme
+  ChannelPreviewTheme copyWith({
+    TextStyle? title,
+    TextStyle? subtitle,
+    TextStyle? lastMessageAt,
+    AvatarTheme? avatarTheme,
+    Color? unreadCounterColor,
+    double? indicatorIconSize,
+  }) =>
+      ChannelPreviewTheme(
+        title: title ?? this.title,
+        subtitle: subtitle ?? this.subtitle,
+        lastMessageAt: lastMessageAt ?? this.lastMessageAt,
+        avatarTheme: avatarTheme ?? this.avatarTheme,
+        unreadCounterColor: unreadCounterColor ?? this.unreadCounterColor,
+        indicatorIconSize: indicatorIconSize ?? this.indicatorIconSize,
+      );
+
+  /// Merge with theme
+  ChannelPreviewTheme merge(ChannelPreviewTheme? other) {
+    if (other == null) return this;
+    return copyWith(
+      title: title?.merge(other.title) ?? other.title,
+      subtitle: subtitle?.merge(other.subtitle) ?? other.subtitle,
+      lastMessageAt:
+      lastMessageAt?.merge(other.lastMessageAt) ?? other.lastMessageAt,
+      avatarTheme: avatarTheme?.merge(other.avatarTheme) ?? other.avatarTheme,
+      unreadCounterColor: other.unreadCounterColor,
+    );
+  }
+}
